Update for current ilch version, validaten and more

+++ controllers/admin/Index.php
@@ -4,6 +4,7 @@ namespace Modules\Twitchstreams\Controllers\Admin;
 
 use \Modules\Twitchstreams\Mappers\Streamer as StreamerMapper;
 use \Modules\Twitchstreams\Models\Streamer as StreamerModel;
+use Ilch\Validation;
 
 class Index extends \Ilch\Controller\Admin
 {
@@ -14,13 +15,13 @@ class Index extends \Ilch\Controller\Admin
                 'name' => 'menuStreamer',
                 'active' => false,
                 'icon' => 'fa fa-th-list',
-                'url' => $this->getLayout()->getUrl(['controller' => 'index', 'action' => 'index'])
-            ],
-            [
-                'name' => 'add',
-                'active' => false,
-                'icon' => 'fa fa-plus-circle',
-                'url' => $this->getLayout()->getUrl(['controller' => 'index', 'action' => 'treat'])
+                'url' => $this->getLayout()->getUrl(['controller' => 'index', 'action' => 'index']),
+                [
+                    'name' => 'add',
+                    'active' => false,
+                    'icon' => 'fa fa-plus-circle',
+                    'url' => $this->getLayout()->getUrl(['controller' => 'index', 'action' => 'treat'])
+                ]
             ],
             [
                 'name' => 'settings',
@@ -28,12 +29,10 @@ class Index extends \Ilch\Controller\Admin
                 'icon' => 'fa fa-cogs',
                 'url' => $this->getLayout()->getUrl(['controller' => 'settings', 'action' => 'index'])
             ]
-        ]; 
+        ];
 
-        if ($this->getRequest()->getControllerName() == 'index' AND $this->getRequest()->getActionName() == 'treat') {
-            $items[1]['active'] = true;
-        } elseif ($this->getRequest()->getControllerName() == 'settings') {
-            $items[2]['active'] = true;
+        if ($this->getRequest()->getActionName() == 'treat') {
+            $items[0][0]['active'] = true;
         } else {
             $items[0]['active'] = true;
         }
@@ -50,8 +49,16 @@ class Index extends \Ilch\Controller\Admin
         $mapper = new StreamerMapper();
 
         $this->getLayout()->getAdminHmenu()
-                ->add($this->getTranslator()->trans('twitchstreams'), ['controller' => 'index', 'action' => 'index'])
-                ->add($this->getTranslator()->trans('menuStreamer'), ['action' => 'index']);
+            ->add($this->getTranslator()->trans('twitchstreams'), ['controller' => 'index', 'action' => 'index'])
+            ->add($this->getTranslator()->trans('menuStreamer'), ['action' => 'index']);
+
+        if ($this->getRequest()->getPost('check_streamer')) {
+            if ($this->getRequest()->getPost('action') == 'delete') {
+                foreach ($this->getRequest()->getPost('check_streamer') as $id) {
+                    $mapper->delete($id);
+                }
+            }
+        }
 
         $this->getView()->set('streamer', $mapper->getStreamer());
     }
@@ -62,23 +69,23 @@ class Index extends \Ilch\Controller\Admin
 
         if ($this->getRequest()->getParam('id')) {
             $this->getLayout()->getAdminHmenu()
-                    ->add($this->getTranslator()->trans('twitchstreams'), ['action' => 'index'])
-                    ->add($this->getTranslator()->trans('edit'), ['action' => 'treat']);
+                ->add($this->getTranslator()->trans('twitchstreams'), ['action' => 'index'])
+                ->add($this->getTranslator()->trans('edit'), ['action' => 'treat']);
 
             $this->getView()->set('streamer', $mapper->readById($this->getRequest()->getParam('id')));
         } else {
             $this->getLayout()->getAdminHmenu()
-                    ->add($this->getTranslator()->trans('twitchstreams'), ['controller' => 'index', 'action' => 'index'])
-                    ->add($this->getTranslator()->trans('add'), ['action' => 'treat']);
+                ->add($this->getTranslator()->trans('twitchstreams'), ['controller' => 'index', 'action' => 'index'])
+                ->add($this->getTranslator()->trans('add'), ['action' => 'treat']);
         }
 
         if ($this->getRequest()->getParam('id')) {
             $streamerArray = $mapper->readById($this->getRequest()->getParam('id'));
             $streamerModel = new StreamerModel();
-            $streamerModel->setId($streamerArray['id']);
-            $streamerModel->setUser($streamerArray['user']);
-            $streamerModel->setOnline($streamerArray['online']);
-            $streamerModel->setGame($streamerArray['game']);
+            $streamerModel->setId($streamerArray['id'])
+                ->setUser($streamerArray['user'])
+                ->setOnline($streamerArray['online'])
+                ->setGame($streamerArray['game']);
 
             $this->getView()->set('streamer', $streamerModel);
         }
@@ -87,25 +94,35 @@ class Index extends \Ilch\Controller\Admin
             $mapper = new StreamerMapper();
             $model = new StreamerModel();
 
-            if ($this->getRequest()->getParam('id')) {
-                $model->setId($this->getRequest()->getParam('id'));
-            }
+            Validation::setCustomFieldAliases([
+                'inputUser' => 'streamer'
+            ]);
 
-            $user = $this->getRequest()->getPost('inputUser');
+            $validation = Validation::create($this->getRequest()->getPost(), [
+                'inputUser' => 'required'
+            ]);
 
-            if (empty($user)) {
-                $this->addMessage('missingUser', 'danger');
-            } else {
-                $model->setUser($user);
-                $model->setOnline(0);
+            if ($validation->isValid()) {
+                if ($this->getRequest()->getParam('id')) {
+                    $model->setId($this->getRequest()->getParam('id'));
+                }
+
+                $model->setUser($this->getRequest()->getPost('inputUser'))
+                    ->setOnline(0);
                 $mapper->save($model);
 
-                $mapper->updateOnlineStreamer();
+                $mapper->updateOnlineStreamer($this->getConfig()->get('twitchstreams_apiKey'));
 
-                $this->addMessage('saveSuccess');
-                
-                $this->redirect(['action' => 'index']);
+                $this->redirect()
+                    ->withMessage('saveSuccess')
+                    ->to(['action' => 'index']);
             }
+
+            $this->addMessage($validation->getErrorBag()->getErrorMessages(), 'danger', true);
+            $this->redirect()
+                ->withInput()
+                ->withErrors($validation->getErrorBag())
+                ->to(['action' => 'treat']);
         }
     }
 
@@ -113,22 +130,27 @@ class Index extends \Ilch\Controller\Admin
     {
         $mapper = new StreamerMapper();
 
-        $mapper->updateOnlineStreamer();
+        $mapper->updateOnlineStreamer($this->getConfig()->get('twitchstreams_apiKey'));
 
-        $this->addMessage('updateSuccess');
-
-        $this->redirect(['action' => 'index']);
+        $this->redirect()
+            ->withMessage('updateSuccess')
+            ->to(['action' => 'index']);
     }
 
     public function deleteAction()
     {
-        if ($this->getRequest()->getParam('id')) {
-            $mapper = new StreamerMapper();
-            $mapper->delete($this->getRequest()->getParam('id'));
+        if ($this->getRequest()->isSecure()) {
+            if ($this->getRequest()->getParam('id')) {
+                $mapper = new StreamerMapper();
+                $mapper->delete($this->getRequest()->getParam('id'));
 
-            $this->addMessage('deleteSuccess');
-
-            $this->redirect(['action' => 'index']);
+                $this->redirect()
+                    ->withMessage('deleteSuccess')
+                    ->to(['action' => 'index']);
+            }
         }
+
+        $this->redirect()
+            ->to(['action' => 'index']);
     }
 }
